refactor(fe): separate hook from compoenent
